refactor: 상속으로 우테코 라이브러리 영향 최소화하기
