+++ src/test/java/lotto/view/LottoBuyInputViewTest.java
@@ -2,6 +2,7 @@ package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import lotto.message.ErrorMessage;
+import lotto.view.input.LottoBuyInputView;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
